pull apart large Root component into parts
