+++ packages/trees/src/components/Root.tsx
@@ -1,6 +1,5 @@
 /** @jsxImportSource preact */
-import { Component, createElement, Fragment } from 'preact';
-import type { FunctionComponent, JSX } from 'preact';
+import type { JSX } from 'preact';
 import {
   useCallback,
   useEffect,
@@ -40,7 +39,6 @@ import {
   type FileTreeCallbacks,
   type FileTreeHandle,
   type FileTreeOptions,
-  type FileTreeSelectionItem,
   type FileTreeStateConfig,
   isRenamingEnabled,
 } from '../FileTree';
@@ -49,23 +47,25 @@ import { generateSyncDataLoaderFromTreeData } from '../loader/sync';
 import type { SVGSpriteNames } from '../sprite';
 import type { FileTreeNode } from '../types';
 import { computeNewFilesAfterDrop } from '../utils/computeNewFilesAfterDrop';
-import { controlledExpandedPathsToExpandedIds } from '../utils/controlledExpandedState';
-import {
-  expandPathsWithAncestors,
-  filterOrphanedPaths,
-} from '../utils/expandPaths';
 import { fileListToTree } from '../utils/fileListToTree';
 import { getGitStatusSignature } from '../utils/getGitStatusSignature';
 import { getSelectionPath } from '../utils/getSelectionPath';
-import {
-  remapExpandedPathsForFolderRename,
-  renameFileTreePaths,
-} from '../utils/renameFileTreePaths';
+import { renameFileTreePaths } from '../utils/renameFileTreePaths';
 import type { ChildrenSortOption } from '../utils/sortChildren';
 import { useContextMenuController } from './hooks/useContextMenuController';
+import {
+  getFilesSignature,
+  type PendingDropTarget,
+  type PendingRenameExpandedRemap,
+  type PendingRenameFocusRestore,
+  useExpansionMigration,
+} from './hooks/useExpansionMigration';
+import { useFlattenedDropTarget } from './hooks/useFlattenedDropTarget';
+import { useStateChangeCallbacks } from './hooks/useStateChangeCallbacks';
 import { useTree } from './hooks/useTree';
+import { useTreeStateConfig } from './hooks/useTreeStateConfig';
 import { Icon } from './Icon';
-import { MiddleTruncate, Truncate } from './OverflowText';
+import { TreeItem } from './TreeItem';
 import { VirtualizedList } from './VirtualizedList';
 
 export interface FileTreeRootProps {
@@ -75,364 +75,7 @@ export interface FileTreeRootProps {
   callbacksRef?: { current: FileTreeCallbacks };
 }
 
-// Local memo implementation to avoid importing from preact/compat, which
-// declares `export as namespace React` and pollutes the global type namespace,
-// breaking the React wrapper's JSX types.
-function memo<P>(
-  c: FunctionComponent<P>,
-  comparer: (prev: P, next: P) => boolean
-): FunctionComponent<P> {
-  class Memoed extends Component<P> {
-    override shouldComponentUpdate(nextProps: P) {
-      return !comparer(this.props as P, nextProps);
-    }
-    override render() {
-      return createElement(
-        c as FunctionComponent,
-        this.props as Record<string, unknown>
-      );
-    }
-  }
-  Memoed.displayName = `Memo(${c.displayName ?? c.name ?? 'Component'})`;
-  return Memoed as unknown as FunctionComponent<P>;
-}
-
-const getFilesSignature = (files: string[]): string =>
-  `${files.length}\0${files.join('\0')}`;
-
 const EMPTY_ANCESTORS: string[] = [];
-
-function resolvePathToRenderedId(
-  path: string,
-  pathToId: Map<string, string>,
-  flattenEmptyDirectories: boolean
-): string | null {
-  if (path.startsWith(FLATTENED_PREFIX)) {
-    return pathToId.get(path) ?? null;
-  }
-  if (flattenEmptyDirectories) {
-    return pathToId.get(FLATTENED_PREFIX + path) ?? pathToId.get(path) ?? null;
-  }
-  return pathToId.get(path) ?? null;
-}
-
-function RenameInput({
-  ariaLabel,
-  isFlattened,
-  renameInputProps,
-}: {
-  ariaLabel: string;
-  isFlattened?: boolean;
-  renameInputProps: Record<string, unknown>;
-}): JSX.Element {
-  return (
-    <input
-      data-item-rename-input
-      {...(isFlattened === true
-        ? { 'data-item-flattened-rename-input': true }
-        : {})}
-      aria-label={ariaLabel}
-      onClick={(e) => e.stopPropagation()}
-      onPointerDown={(e) => e.stopPropagation()}
-      {...renameInputProps}
-    />
-  );
-}
-
-function FlattenedDirectoryName({
-  tree,
-  idToPath,
-  flattens,
-  fallbackName,
-  renameInputProps,
-}: {
-  tree: TreeInstance<FileTreeNode>;
-  idToPath: Map<string, string>;
-  flattens: string[];
-  fallbackName: string;
-  renameInputProps?: Record<string, unknown> | null;
-}): JSX.Element {
-  'use no memo';
-  const segments = useMemo(() => {
-    const result: { id: string; label: string }[] = [];
-    for (const id of flattens) {
-      const item = tree.getItemInstance(id);
-      if (item != null) {
-        result.push({ id: item.getId(), label: item.getItemName() });
-      } else {
-        const path = idToPath.get(id);
-        const label = path != null ? (path.split('/').pop() ?? id) : id;
-        result.push({ id, label });
-      }
-    }
-    return result;
-  }, [flattens, tree, idToPath]);
-
-  if (segments.length === 0) {
-    if (renameInputProps != null) {
-      return (
-        <RenameInput
-          ariaLabel={`Rename ${fallbackName}`}
-          isFlattened
-          renameInputProps={renameInputProps}
-        />
-      );
-    }
-    return (
-      <span data-item-flattened-subitems>
-        {fallbackName.replace(/\//g, ' / ')}
-      </span>
-    );
-  }
-
-  return (
-    <span data-item-flattened-subitems>
-      {segments.map(({ id, label }, index) => {
-        const isLast = index === segments.length - 1;
-        return (
-          <Fragment key={id}>
-            <span data-item-flattened-subitem={id}>
-              {isLast && renameInputProps != null ? (
-                <RenameInput
-                  ariaLabel={`Rename ${label}`}
-                  isFlattened
-                  renameInputProps={renameInputProps}
-                />
-              ) : (
-                <Truncate>{label}</Truncate>
-              )}
-            </span>
-            {!isLast ? ' / ' : ''}
-          </Fragment>
-        );
-      })}
-    </span>
-  );
-}
-
-interface TreeItemProps {
-  item: ItemInstance<FileTreeNode>;
-  tree: TreeInstance<FileTreeNode>;
-  itemId: string;
-  hasChildren: boolean;
-  isExpanded: boolean;
-  itemName: string;
-  level: number;
-  isSelected: boolean;
-  isFocused: boolean;
-  isSearchMatch: boolean;
-  isDragTarget: boolean;
-  isDragging: boolean;
-  isDnD: boolean;
-  isRenaming: boolean;
-  isFlattenedDirectory: boolean;
-  isLocked: boolean;
-  gitStatus: string | undefined;
-  containsGitChange: boolean;
-  flattens: string[] | undefined;
-  idToPath: Map<string, string>;
-  ancestors: string[];
-  treeDomId: string;
-  remapIcon: (name: SVGSpriteNames) => {
-    name: string;
-    remappedFrom?: string;
-    width?: number;
-    height?: number;
-    viewBox?: string;
-  };
-  detectFlattenedSubfolder: (e: DragEvent) => void;
-  clearFlattenedSubfolder: () => void;
-}
-
-function treeItemPropsAreEqual(
-  prev: Readonly<TreeItemProps>,
-  next: Readonly<TreeItemProps>
-): boolean {
-  return (
-    prev.itemId === next.itemId &&
-    prev.hasChildren === next.hasChildren &&
-    prev.isExpanded === next.isExpanded &&
-    prev.itemName === next.itemName &&
-    prev.level === next.level &&
-    prev.isSelected === next.isSelected &&
-    prev.isFocused === next.isFocused &&
-    prev.isSearchMatch === next.isSearchMatch &&
-    prev.isDragTarget === next.isDragTarget &&
-    prev.isDragging === next.isDragging &&
-    prev.isDnD === next.isDnD &&
-    prev.isRenaming === next.isRenaming &&
-    prev.isFlattenedDirectory === next.isFlattenedDirectory &&
-    prev.isLocked === next.isLocked &&
-    prev.gitStatus === next.gitStatus &&
-    prev.containsGitChange === next.containsGitChange &&
-    prev.flattens === next.flattens &&
-    prev.ancestors === next.ancestors &&
-    prev.treeDomId === next.treeDomId
-  );
-}
-
-function TreeItemInner({
-  item,
-  tree,
-  itemId,
-  hasChildren,
-  itemName,
-  level,
-  isSelected,
-  isFocused,
-  isSearchMatch,
-  isDragTarget,
-  isDragging,
-  isDnD,
-  isRenaming,
-  isFlattenedDirectory,
-  isLocked,
-  gitStatus: itemGitStatus,
-  containsGitChange: itemContainsGitChange,
-  flattens,
-  idToPath,
-  ancestors,
-  treeDomId,
-  remapIcon,
-  detectFlattenedSubfolder,
-  clearFlattenedSubfolder,
-}: TreeItemProps): JSX.Element {
-  'use no memo';
-  const startWithCapital =
-    itemName.charAt(0).toUpperCase() === itemName.charAt(0);
-  const alignCapitals = startWithCapital;
-
-  const selectionProps = isSelected ? { 'data-item-selected': true } : {};
-  const focusedProps = isFocused ? { 'data-item-focused': true } : {};
-  const searchMatchProps = isSearchMatch
-    ? { 'data-item-search-match': true }
-    : {};
-  const dragProps = isDnD
-    ? {
-        ...(isDragTarget && { 'data-item-drag-target': true }),
-        ...(isDragging && { 'data-item-dragging': true }),
-      }
-    : {};
-  const gitStatusProps = {
-    ...(itemGitStatus != null && {
-      'data-item-git-status': itemGitStatus,
-    }),
-    ...(itemContainsGitChange && {
-      'data-item-contains-git-change': 'true',
-    }),
-  };
-
-  const baseProps = item.getProps();
-  const itemProps =
-    isDnD && isFlattenedDirectory
-      ? {
-          ...baseProps,
-          onDragOver: (e: DragEvent) => {
-            (baseProps.onDragOver as ((e: DragEvent) => void) | undefined)?.(e);
-            detectFlattenedSubfolder(e);
-          },
-          onDragLeave: (e: DragEvent) => {
-            clearFlattenedSubfolder();
-            (baseProps.onDragLeave as ((e: DragEvent) => void) | undefined)?.(
-              e
-            );
-          },
-          onDrop: (e: DragEvent) => {
-            (baseProps.onDrop as ((e: DragEvent) => void) | undefined)?.(e);
-            clearFlattenedSubfolder();
-          },
-        }
-      : baseProps;
-  const statusLabel =
-    itemGitStatus === 'added'
-      ? 'A'
-      : itemGitStatus === 'deleted'
-        ? 'D'
-        : itemGitStatus === 'modified'
-          ? 'M'
-          : null;
-  const showStatusDot = statusLabel == null && itemContainsGitChange;
-  const renameInputProps =
-    isRenaming && item.getRenameInputProps != null
-      ? item.getRenameInputProps()
-      : null;
-
-  const getItemDomId = (id: string) => `${treeDomId}-${id}`;
-
-  return (
-    <button
-      data-type="item"
-      data-item-type={hasChildren ? 'folder' : 'file'}
-      aria-label={itemName}
-      {...selectionProps}
-      {...searchMatchProps}
-      {...focusedProps}
-      {...dragProps}
-      {...gitStatusProps}
-      data-item-id={itemId}
-      id={getItemDomId(itemId)}
-      {...itemProps}
-      key={itemId}
-    >
-      {level > 0 ? (
-        <div data-item-section="spacing">
-          {Array.from({ length: level }).map((_, index) => (
-            <div
-              key={index}
-              data-item-section="spacing-item"
-              data-ancestor-id={ancestors[index]}
-            />
-          ))}
-        </div>
-      ) : null}
-      <div data-item-section="icon">
-        {hasChildren ? (
-          <Icon
-            {...remapIcon('file-tree-icon-chevron')}
-            alignCapitals={alignCapitals}
-          />
-        ) : (
-          <Icon {...remapIcon('file-tree-icon-file')} />
-        )}
-      </div>
-      <div data-item-section="content">
-        {isFlattenedDirectory ? (
-          <FlattenedDirectoryName
-            tree={tree}
-            idToPath={idToPath}
-            flattens={flattens ?? []}
-            fallbackName={itemName}
-            renameInputProps={renameInputProps}
-          />
-        ) : renameInputProps != null ? (
-          <RenameInput
-            ariaLabel={`Rename ${itemName}`}
-            renameInputProps={renameInputProps}
-          />
-        ) : (
-          <MiddleTruncate minimumLength={5} split="extension">
-            {itemName}
-          </MiddleTruncate>
-        )}
-      </div>
-
-      {statusLabel || showStatusDot ? (
-        <div data-item-section="status">
-          {statusLabel ?? (
-            <Icon {...remapIcon('file-tree-icon-dot')} width={6} height={6} />
-          )}
-        </div>
-      ) : null}
-      {isLocked ? (
-        <div data-item-section="lock">
-          <Icon {...remapIcon('file-tree-icon-lock')} />
-        </div>
-      ) : null}
-    </button>
-  );
-}
-
-const TreeItem = memo(TreeItemInner, treeItemPropsAreEqual);
 
 export function Root({
   fileTreeOptions,
@@ -494,9 +137,9 @@ export function Root({
   );
 
   // Resolve sort option to a comparator (or undefined for default behavior).
-  // `false` → preserve insertion order and skip sort work.
-  // `{ comparator }` → custom comparator.
-  // `true` / `undefined` → undefined (use default).
+  // `false` -> preserve insertion order and skip sort work.
+  // `{ comparator }` -> custom comparator.
+  // `true` / `undefined` -> undefined (use default).
   const sortComparator = useMemo<ChildrenSortOption | undefined>(
     () =>
       sortOption === false
@@ -512,7 +155,7 @@ export function Root({
     [files, sortComparator]
   );
 
-  // Build path↔id maps from treeData
+  // Build path<->id maps from treeData
   const { pathToId, idToPath } = useMemo(() => {
     const p2i = new Map<string, string>();
     const i2p = new Map<string, string>();
@@ -530,197 +173,13 @@ export function Root({
     ancestorChainsCacheRef.current.clear();
   }
 
-  const restTreeConfig = useMemo(() => {
-    const mapId = (item: string): string => {
-      if (treeData[item] != null) {
-        return item;
-      }
-      return pathToId.get(item) ?? item;
-    };
+  const restTreeConfig = useTreeStateConfig({
+    treeData,
+    pathToId,
+    stateConfig,
+    flattenEmptyDirectories,
+  });
 
-    const mapIds = (items: string[] | undefined): string[] | undefined => {
-      if (items == null) {
-        return undefined;
-      }
-      let changed = false;
-      const mapped = items.map((item) => {
-        const mappedItem = mapId(item);
-        if (mappedItem !== item) {
-          changed = true;
-        }
-        return mappedItem;
-      });
-      return changed ? mapped : items;
-    };
-
-    type TreeStateConfig = {
-      expandedItems?: string[];
-      selectedItems?: string[];
-      focusedItem?: string | null;
-      renamingItem?: string | null;
-      checkedItems?: string[];
-      loadingCheckPropagationItems?: string[];
-      [key: string]: unknown;
-    };
-
-    const mapState = (state: TreeStateConfig | undefined) => {
-      if (state == null) {
-        return { state, changed: false };
-      }
-      let changed = false;
-      const nextState: TreeStateConfig = { ...state };
-
-      const mappedExpanded = mapIds(state.expandedItems);
-      if (mappedExpanded !== state.expandedItems) {
-        nextState.expandedItems = mappedExpanded;
-        changed = true;
-      }
-
-      const mappedSelected = mapIds(state.selectedItems);
-      if (mappedSelected !== state.selectedItems) {
-        nextState.selectedItems = mappedSelected;
-        changed = true;
-      }
-
-      const mappedFocused =
-        state.focusedItem != null
-          ? mapId(state.focusedItem)
-          : state.focusedItem;
-      if (mappedFocused !== state.focusedItem) {
-        nextState.focusedItem = mappedFocused;
-        changed = true;
-      }
-
-      const mappedRenaming =
-        state.renamingItem != null
-          ? mapId(state.renamingItem)
-          : state.renamingItem;
-      if (mappedRenaming !== state.renamingItem) {
-        nextState.renamingItem = mappedRenaming;
-        changed = true;
-      }
-
-      const mappedChecked = mapIds(state.checkedItems);
-      if (mappedChecked !== state.checkedItems) {
-        nextState.checkedItems = mappedChecked;
-        changed = true;
-      }
-
-      const mappedLoadingChecked = mapIds(state.loadingCheckPropagationItems);
-      if (mappedLoadingChecked !== state.loadingCheckPropagationItems) {
-        nextState.loadingCheckPropagationItems = mappedLoadingChecked;
-        changed = true;
-      }
-
-      return { state: changed ? nextState : state, changed };
-    };
-
-    const baseConfig: TreeStateConfig = {};
-
-    const mapPathToId = (path: string): string | undefined => {
-      // If the caller explicitly passes a flattened path, respect it.
-      if (path.startsWith(FLATTENED_PREFIX)) {
-        return pathToId.get(path);
-      }
-
-      const shouldFlatten = flattenEmptyDirectories === true;
-
-      // Only prefer flattened IDs when the tree is actually rendering flattened
-      // directories. Otherwise, selecting a path that *could* be flattened would
-      // target a hidden node and the visible folder would not be marked selected.
-      if (shouldFlatten) {
-        return pathToId.get(FLATTENED_PREFIX + path) ?? pathToId.get(path);
-      }
-      return pathToId.get(path);
-    };
-
-    const mapPathsToIds = (
-      paths: string[] | undefined
-    ): string[] | undefined => {
-      if (paths == null) return undefined;
-      const ids = paths
-        .map(mapPathToId)
-        .filter((id): id is string => id != null);
-      return ids.length > 0 ? ids : [];
-    };
-
-    // Merge top-level initialExpandedItems/initialSelectedItems/initialSearchQuery into config.initialState
-    const topLevelInitialExpanded = stateConfig?.initialExpandedItems;
-    const topLevelInitialSelected = stateConfig?.initialSelectedItems;
-    const topLevelInitialSearch = stateConfig?.initialSearchQuery;
-    const topLevelInitialExpandedIds =
-      topLevelInitialExpanded != null
-        ? expandPathsWithAncestors(topLevelInitialExpanded, pathToId, {
-            flattenEmptyDirectories,
-          })
-        : undefined;
-    const topLevelInitialSelectedIds = mapPathsToIds(topLevelInitialSelected);
-    const hasTopLevelInitial =
-      topLevelInitialExpanded != null ||
-      topLevelInitialSelected != null ||
-      topLevelInitialSearch != null;
-
-    const mergedInitialState = hasTopLevelInitial
-      ? {
-          ...(baseConfig.initialState as TreeStateConfig | undefined),
-          ...(topLevelInitialExpandedIds != null && {
-            expandedItems: topLevelInitialExpandedIds,
-          }),
-          ...(topLevelInitialSelectedIds != null && {
-            selectedItems: topLevelInitialSelectedIds,
-          }),
-          ...(topLevelInitialSearch != null && {
-            search: topLevelInitialSearch,
-          }),
-        }
-      : (baseConfig.initialState as TreeStateConfig | undefined);
-
-    // Merge top-level expandedItems/selectedItems into config.state
-    const topLevelExpanded = stateConfig?.expandedItems;
-    const topLevelSelected = stateConfig?.selectedItems;
-    const topLevelExpandedIds =
-      topLevelExpanded != null
-        ? controlledExpandedPathsToExpandedIds(topLevelExpanded, pathToId, {
-            flattenEmptyDirectories,
-          })
-        : undefined;
-    const topLevelSelectedIds = mapPathsToIds(topLevelSelected);
-    const hasTopLevelState =
-      topLevelExpanded != null || topLevelSelected != null;
-
-    const mergedState = hasTopLevelState
-      ? {
-          ...(baseConfig.state as TreeStateConfig | undefined),
-          ...(topLevelExpandedIds != null && {
-            expandedItems: topLevelExpandedIds,
-          }),
-          ...(topLevelSelectedIds != null && {
-            selectedItems: topLevelSelectedIds,
-          }),
-        }
-      : (baseConfig.state as TreeStateConfig | undefined);
-
-    const configWithMergedState = {
-      ...baseConfig,
-      ...(mergedInitialState != null && { initialState: mergedInitialState }),
-      ...(mergedState != null && { state: mergedState }),
-    };
-
-    const initialState = mapState(
-      configWithMergedState.initialState as TreeStateConfig
-    );
-    const state = mapState(configWithMergedState.state as TreeStateConfig);
-
-    if (!initialState.changed && !state.changed) {
-      return configWithMergedState;
-    }
-
-    return {
-      ...configWithMergedState,
-      ...(initialState.state != null && { initialState: initialState.state }),
-      ...(state.state != null && { state: state.state }),
-    };
-  }, [treeData, pathToId, stateConfig, flattenEmptyDirectories]);
   const dataLoader = useMemo(
     () =>
       useLazyDataLoader === true
@@ -770,107 +229,26 @@ export function Root({
   const filesRef = useRef(files);
   filesRef.current = files;
 
+  // Ref populated after useTree — allows hooks called before useTree to
+  // access the tree instance at event-handler time (not render time).
+  const treeRef = useRef<TreeInstance<FileTreeNode> | null>(null);
+
   // --- Flattened sub-folder drop targeting ---
-  const flattenedDropSubfolderIdRef = useRef<string | null>(null);
-  const flattenedHighlightRef = useRef<HTMLElement | null>(null);
+  const {
+    flattenedDropSubfolderIdRef,
+    detectFlattenedSubfolder,
+    clearFlattenedSubfolder,
+    detectFlattenedSubfolderFromPoint,
+  } = useFlattenedDropTarget(treeRef);
 
-  const detectFlattenedSubfolder = useCallback((e: DragEvent) => {
-    let el = e.target as HTMLElement | null;
-    if (el != null && el.nodeType === Node.TEXT_NODE) {
-      el = el.parentElement;
-    }
-    const span = el?.closest?.(
-      '[data-item-flattened-subitem]'
-    ) as HTMLElement | null;
-    const id = span?.getAttribute('data-item-flattened-subitem') ?? null;
-
-    if (id === flattenedDropSubfolderIdRef.current) return;
-
-    if (flattenedHighlightRef.current != null) {
-      flattenedHighlightRef.current.removeAttribute(
-        'data-item-flattened-subitem-drag-target'
-      );
-    }
-
-    if (span != null && id != null) {
-      span.setAttribute('data-item-flattened-subitem-drag-target', 'true');
-      flattenedHighlightRef.current = span;
-      flattenedDropSubfolderIdRef.current = id;
-    } else {
-      flattenedHighlightRef.current = null;
-      flattenedDropSubfolderIdRef.current = null;
-    }
-  }, []);
-
-  const clearFlattenedSubfolder = useCallback(() => {
-    if (flattenedHighlightRef.current != null) {
-      flattenedHighlightRef.current.removeAttribute(
-        'data-item-flattened-subitem-drag-target'
-      );
-    }
-    flattenedHighlightRef.current = null;
-    flattenedDropSubfolderIdRef.current = null;
-  }, []);
-
-  const detectFlattenedSubfolderFromPoint = useCallback(
-    (clientX: number, clientY: number) => {
-      const treeEl = tree.getElement();
-      const root = treeEl?.getRootNode() as Document | ShadowRoot;
-      let el = (root ?? document).elementFromPoint(
-        clientX,
-        clientY
-      ) as HTMLElement | null;
-      if (el != null && el.nodeType === Node.TEXT_NODE) {
-        el = el.parentElement;
-      }
-      const span = el?.closest?.(
-        '[data-item-flattened-subitem]'
-      ) as HTMLElement | null;
-      const id = span?.getAttribute('data-item-flattened-subitem') ?? null;
-
-      if (id === flattenedDropSubfolderIdRef.current) return;
-
-      if (flattenedHighlightRef.current != null) {
-        flattenedHighlightRef.current.removeAttribute(
-          'data-item-flattened-subitem-drag-target'
-        );
-      }
-
-      if (span != null && id != null) {
-        span.setAttribute('data-item-flattened-subitem-drag-target', 'true');
-        flattenedHighlightRef.current = span;
-        flattenedDropSubfolderIdRef.current = id;
-      } else {
-        flattenedHighlightRef.current = null;
-        flattenedDropSubfolderIdRef.current = null;
-      }
-    },
-    // tree.getElement() is stable across renders
-    // oxlint-disable-next-line react-hooks/exhaustive-deps
-    []
+  // Pending refs for expansion migration — owned here so both the
+  // onDrop/onRename callbacks and the migration hook can access them.
+  const pendingDropTargetExpandRef = useRef<PendingDropTarget | null>(null);
+  const pendingRenameExpandedRemapRef =
+    useRef<PendingRenameExpandedRemap | null>(null);
+  const pendingRenameFocusRestoreRef = useRef<PendingRenameFocusRestore | null>(
+    null
   );
-
-  // Keep the previous idToPath so we can translate stale expanded IDs → paths
-  // when files change (DnD or controlled update).
-  const prevIdToPathRef = useRef<Map<string, string>>(idToPath);
-  // DnD-only: pending drop target to auto-expand if/when the exact drop result
-  // is applied to files.
-  const pendingDropTargetExpandRef = useRef<{
-    path: string;
-    expectedFilesSignature: string;
-  } | null>(null);
-  // Rename-only: stores the old/new folder path so the expand-migration
-  // effect can remap expanded children once the new idToPath is available.
-  const pendingRenameExpandedRemapRef = useRef<{
-    sourcePath: string;
-    destinationPath: string;
-  } | null>(null);
-  // Rename-only: stores the renamed destination path so focus can be moved
-  // to the remapped item ID once files/pathToId have updated.
-  const pendingRenameFocusRestoreRef = useRef<{
-    destinationPath: string;
-    expectedFilesSignature: string;
-  } | null>(null);
 
   const onDropHandler = useCallback(
     (
@@ -912,7 +290,7 @@ export function Root({
 
       callbacksRef?.current._onDragMoveFiles?.(newFiles);
     },
-    [callbacksRef, onCollision, idToPath]
+    [callbacksRef, onCollision, idToPath, flattenedDropSubfolderIdRef]
   );
 
   // Track search state via ref so the canDrag callback (evaluated at event
@@ -920,8 +298,8 @@ export function Root({
   const searchActiveRef = useRef(false);
 
   // Search config is read by fileTreeSearchFeature via getConfig().
-  // via getConfig(). We spread it from a variable to bypass excess property
-  // checks on the TreeConfig object literal.
+  // We spread it from a variable to bypass excess property checks on the
+  // TreeConfig object literal.
   const searchModeConfig = { fileTreeSearchMode, search };
   const gitStatusConfig = {
     gitStatus,
@@ -941,6 +319,7 @@ export function Root({
     contextMenuEnabled: isContextMenuEnabled,
     onContextMenuRequest: handleContextMenuFeatureRequest,
   };
+
   const tree = useTree<FileTreeNode>({
     ...restTreeConfig,
     ...searchModeConfig,
@@ -1047,6 +426,19 @@ export function Root({
       _onTouchDragEnd: clearFlattenedSubfolder,
     }),
   });
+  treeRef.current = tree;
+
+  // --- Expansion migration ---
+  useExpansionMigration({
+    tree,
+    files,
+    pathToId,
+    idToPath,
+    flattenEmptyDirectories,
+    pendingDropTargetExpandRef,
+    pendingRenameExpandedRemapRef,
+    pendingRenameFocusRestoreRef,
+  });
 
   const getAncestors = useCallback(
     (itemId: string): string[] => {
@@ -1100,36 +492,6 @@ export function Root({
   const focusedItemId = tree.getState().focusedItem ?? null;
   const hasFocusedItem = focusedItemId != null;
 
-  // Detect stale expanded IDs when the file list changes. Flattened chains
-  // may break or form, causing node IDs to change. We snapshot the expanded
-  // paths using the OLD idToPath so the effect can re-map them to new IDs.
-  // This covers both DnD drops and controlled file updates.
-  const pendingExpandMigrationRef = useRef<string[] | null>(null);
-  if (prevIdToPathRef.current !== idToPath) {
-    const currentExpandedIds = tree.getState().expandedItems ?? [];
-    const hasStaleIds = currentExpandedIds.some(
-      (id: string) => !idToPath.has(id)
-    );
-    if (hasStaleIds) {
-      const oldIdToPath = prevIdToPathRef.current;
-      pendingExpandMigrationRef.current = currentExpandedIds
-        .map((id: string) => oldIdToPath.get(id))
-        .filter((p): p is string => p != null)
-        .map((p: string) =>
-          p.startsWith(FLATTENED_PREFIX) ? p.slice(FLATTENED_PREFIX.length) : p
-        );
-      const pendingRename = pendingRenameExpandedRemapRef.current;
-      if (pendingRename != null) {
-        pendingExpandMigrationRef.current = remapExpandedPathsForFolderRename({
-          expandedPaths: pendingExpandMigrationRef.current,
-          sourcePath: pendingRename.sourcePath,
-          destinationPath: pendingRename.destinationPath,
-        });
-      }
-    }
-  }
-  prevIdToPathRef.current = idToPath;
-
   // Populate handleRef so the FileTree class can call tree methods directly
   useEffect(() => {
     if (handleRef == null) return;
@@ -1144,182 +506,17 @@ export function Root({
     };
   }, [closeContextMenu, tree, pathToId, idToPath, handleRef]);
 
-  // --- Migrate expanded state after file list changes ---
-  // When the file list changes (DnD drop or controlled update), flattened
-  // chains may break or form, changing node IDs. This effect re-maps the
-  // previously-expanded paths to new IDs and optionally expands a drop target
-  // when the applied files match a pending drop result.
-  useEffect(() => {
-    const filesSignature = getFilesSignature(files);
-    const previousPaths = pendingExpandMigrationRef.current;
-    const pendingDropTarget = pendingDropTargetExpandRef.current;
-    const pendingRenameFocus = pendingRenameFocusRestoreRef.current;
-    const dropTarget =
-      pendingDropTarget != null &&
-      pendingDropTarget.expectedFilesSignature === filesSignature
-        ? pendingDropTarget.path
-        : null;
-    const renameFocusPath =
-      pendingRenameFocus != null &&
-      pendingRenameFocus.expectedFilesSignature === filesSignature
-        ? pendingRenameFocus.destinationPath
-        : null;
-    pendingExpandMigrationRef.current = null;
-    pendingDropTargetExpandRef.current = null;
-    pendingRenameExpandedRemapRef.current = null;
-    if (renameFocusPath != null) {
-      pendingRenameFocusRestoreRef.current = null;
-    }
-
-    if (
-      previousPaths == null &&
-      dropTarget == null &&
-      renameFocusPath == null
-    ) {
-      return;
-    }
-
-    if (previousPaths != null || dropTarget != null) {
-      const pathsToExpand = previousPaths != null ? [...previousPaths] : [];
-      if (dropTarget != null) {
-        pathsToExpand.push(dropTarget);
-      }
-
-      const expandIds = expandPathsWithAncestors(pathsToExpand, pathToId, {
-        flattenEmptyDirectories,
-      });
-
-      if (previousPaths != null) {
-        // Full replacement — re-map all expanded paths to new IDs.
-        tree.applySubStateUpdate('expandedItems', () => expandIds);
-      } else {
-        // Just adding the drop target — merge with existing expanded state.
-        const currentExpanded = tree.getState().expandedItems ?? [];
-        const currentSet = new Set(currentExpanded);
-        const newIds = expandIds.filter((id) => !currentSet.has(id));
-        if (newIds.length > 0) {
-          tree.applySubStateUpdate('expandedItems', (prev) => [
-            ...(prev ?? []),
-            ...newIds,
-          ]);
-        }
-      }
-      tree.rebuildTree();
-    }
-
-    if (renameFocusPath != null) {
-      const nextFocusedId = resolvePathToRenderedId(
-        renameFocusPath,
-        pathToId,
-        flattenEmptyDirectories === true
-      );
-      if (nextFocusedId != null) {
-        tree.applySubStateUpdate('focusedItem', nextFocusedId);
-        tree.updateDomFocus();
-      }
-    }
-  }, [files, pathToId, tree, flattenEmptyDirectories]);
-
-  // --- Selection change callback ---
-  const selectionSnapshotRef = useRef<string | null>(null);
-  const selectionSnapshot = tree.getState().selectedItems?.join('|') ?? '';
-
-  useEffect(() => {
-    const onSelection = callbacksRef?.current.onSelection;
-    if (onSelection == null) {
-      return;
-    }
-    if (selectionSnapshotRef.current == null) {
-      selectionSnapshotRef.current = selectionSnapshot;
-      return;
-    }
-    if (selectionSnapshotRef.current === selectionSnapshot) {
-      return;
-    }
-
-    selectionSnapshotRef.current = selectionSnapshot;
-    const selection: FileTreeSelectionItem[] = tree
-      .getSelectedItems()
-      .map((item) => {
-        const data = item.getItemData();
-        return {
-          path: getSelectionPath(data.path),
-          isFolder: data.children?.direct != null,
-        };
-      });
-    onSelection(selection);
-  }, [selectionSnapshot, callbacksRef, tree]);
-
-  // --- Expanded items change callback ---
-  const expandedSnapshotRef = useRef<string | null>(null);
-  const expandedSnapshot = tree.getState().expandedItems?.join('|') ?? '';
-
-  useEffect(() => {
-    const onExpandedItemsChange = callbacksRef?.current.onExpandedItemsChange;
-    if (onExpandedItemsChange == null) {
-      return;
-    }
-    if (expandedSnapshotRef.current == null) {
-      expandedSnapshotRef.current = expandedSnapshot;
-      return;
-    }
-    if (expandedSnapshotRef.current === expandedSnapshot) {
-      return;
-    }
-
-    expandedSnapshotRef.current = expandedSnapshot;
-    const ids = tree.getState().expandedItems ?? [];
-    const paths = [
-      ...new Set(
-        ids
-          .map((id) => idToPath.get(id))
-          .filter((path): path is string => path != null)
-          .map(getSelectionPath)
-      ),
-    ];
-    const effectivePaths = filterOrphanedPaths(
-      paths,
-      pathToId,
-      flattenEmptyDirectories
-    );
-    onExpandedItemsChange(effectivePaths);
-  }, [
-    expandedSnapshot,
-    callbacksRef,
+  const { selectionSnapshot } = useStateChangeCallbacks({
     tree,
+    callbacksRef,
     idToPath,
     pathToId,
     flattenEmptyDirectories,
-  ]);
+  });
 
-  // --- Selected items change callback ---
-  const selectedSnapshotRef = useRef<string | null>(null);
-  const selectedSnapshot = tree.getState().selectedItems?.join('|') ?? '';
-
-  useEffect(() => {
-    const onSelectedItemsChange = callbacksRef?.current.onSelectedItemsChange;
-    if (onSelectedItemsChange == null) {
-      return;
-    }
-    if (selectedSnapshotRef.current == null) {
-      selectedSnapshotRef.current = selectedSnapshot;
-      return;
-    }
-    if (selectedSnapshotRef.current === selectedSnapshot) {
-      return;
-    }
-
-    selectedSnapshotRef.current = selectedSnapshot;
-    const ids = tree.getState().selectedItems ?? [];
-    const paths = ids
-      .map((id) => idToPath.get(id))
-      .filter((path): path is string => path != null)
-      .map(getSelectionPath);
-    onSelectedItemsChange(paths);
-  }, [selectedSnapshot, callbacksRef, tree, idToPath]);
-
-  // When tree mounts with initial search in state, run setSearch once so expand/collapse filter is applied.
-  // useLayoutEffect ensures this runs before paint so the first frame shows the correct expansion.
+  // When tree mounts with initial search in state, run setSearch once so
+  // expand/collapse filter is applied. useLayoutEffect ensures this runs
+  // before paint so the first frame shows the correct expansion.
   useLayoutEffect(() => {
     const search = tree.getState().search;
     if (search != null && search.length > 0) {
@@ -1343,6 +540,7 @@ export function Root({
     }),
     onInput: onChange,
   };
+
   // --- Dynamic guide-line highlighting for selected items ---
   const guideStyleText = useMemo(() => {
     const selectedIds = tree.getState().selectedItems ?? [];
